requirements.txt
asgiref==3.9.2
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
Faker==37.11.0
inflection==0.5.1
Markdown==3.9
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
validate_docbr==1.11.1
